CI: Fix macOS build by working around bogus SDK detection
For whatever reason CMake seems to infer "12.6" as SDK version, because
the GitHub Action macOS runner is based on macOS 12.6. But that number
doesn't match the SDK version, which should be 13.1 here.

This seems to be a long running issue with GitHub Actions' macOS runner:
https://github.com/actions/runner-images/issues/2211

Though I can't tell if it's a CMake bug or a GitHub Actions image setup
issue.